\name{DifferentialExpression}

\alias{miRNATwoGroupDifferentialExpression}

\title{Analyses of differential expression}

\description{These functions assess differential expression for
  particular data types.}

\usage{
miRNATwoGroupDifferentialExpression(targets)
}

\arguments{

  \item{targets}{An object of class \code{Targets} containing slots
    \code{FileName} and \code{Condition}. \code{FileName} is a character
    vector of the URL or other locations of agilent GPR files to be
    used. \code{Condition} is a factor containing two levels, or a
    character vector with two distinct values.}

}

\value{A table generated by calling \pkg{limma} function
  \code{topTable}, summarizing statistics of differential expression for
  each probe. The object is returned as an object of class
  \code{MicroRNATopTable}.}

\examples{
## require(iSalmon)
filenames = source(system.file("script","files.R",
                               package="iSalmon"))[[1]]
conds = c("A","A","B","B")
targets <- data.frame(Condition=conds,
                      FileName=filenames,
                      stringsAsFactors=FALSE)

##From R
topTable <-
    miRNATwoGroupDifferentialExpression(targets)
head(topTable)

##############################################################
## From web service
## targets <- new("Targets", FileName=as.character(filenames),
##                Condition=as.character(conds))
## topTable <-
##     camiRNATwoGroupDifferentialExpression(targets)
## head(topTable)
###############################################################
## system.file("script",
##             "miRNATwoGroupDifferentialExpression.R",
##             package="iSalmon")
}

