\name{TCGAmiRNATwoGroupDifferentialExpression}

\alias{TCGAmiRNATwoGroupDifferentialExpression}

\title{Analyses of differential expression}

\description{This function repeats analysis done in limma for TCGA data.}

\usage{
TCGAmiRNATwoGroupDifferentialExpression(targets)
}

\arguments{

  \item{targets}{An object of class \code{Targets} containing slots
    \code{FileName} and \code{Condition}. \code{FileName} is a character
    vector of the URL or other locations of agilent GPR files to be
    used. \code{Condition} is a factor containing two levels, or a
    character vector with two distinct values.}

}

\value{A table generated by calling \pkg{limma} function
  \code{topTable}, summarizing statistics of differential expression for
  each probe. The object is returned as an object of class
  \code{MicroRNATopTable}.}

\examples{
\dontrun{
require(Rlabkey)
s <- getSession(baseUrl="https://isalmon.fhcrc.org/labkey",
                folderPath="/home/SharedData")
schema <- getSchema(s, "lists")
sample <- schema[["Ovarian_SamplesToUploadedFiles"]]
##lkups <- getLookups(s, sample[["BCRSAMPLEBARCODE"]])
cols <- c(names(sample))
df <- getRows(s, sample, colSelect=cols)

subListSource <- system.file("extdata",
                             "MUTaa_WILD_rem_mod_target_file",
                              package="iSalmon")
subList <- read.table(subListSource, sep="\t", header=TRUE)

final <- merge(subList, df)

targets = data.frame(FileName=final[,5],
                     Condition=final[,3],
                     stringsAsFactors=FALSE)

require(iSalmon)
topTable <-
    TCGAmiRNATwoGroupDifferentialExpression(targets)

head(topTable)
}
}

