\name{TopTableSingleTissueSingleMutgene}

\alias{TopTableSingleTissueSingleMutgene}

\title{Differential expression in samples with known mutation status and tissue
  type}

\description{This helper function tests specific kinds of tissue samples
  with known gene mutations as determined by Sanger sequencing for
  differential expression and reports the results into output files.}

\usage{
TopTableSingleTissueSingleMutgene(miEset,input.tissue,mut.gene, outDir=".")
}
\arguments{

  \item{miEset}{An expression set object based on microRNA data as
    generated by using \code{SangerMiRNAGenerateESet} }
  \item{input.tissue}{The tissue type to check}
  \item{mut.gene}{The mutant gene to check}
  \item{outDir}{The directory where output files will be written.}

}

\value{A directory containing files that indicate which genes if any
  were differentially expressed.  Other output files indicate when a test was
  not permitted and why.}

\examples{
\dontrun{
# 1st we need an eset object
miEset <- SangerMiRNAGenerateESet()
# This example will produce an interesting result
input.tissue = "stomach_non_small_cell_cancer"
mut.gene = "pik3ca"
TopTableSingleTissueSingleMutgene(miEset,input.tissue,mut.gene)


# This one does not meet testing criteria
input.tissue = "stomach_non_small_cell_cancer"
mut.gene = "apc"
TopTableSingleTissueSingleMutgene(miEset,input.tissue,mut.gene)
}
}

